Oops. 9 hours searching and here it is!

+++ src/main/java/com/github/dirtpowered/betaprotocollib/packet/WindowClickPacket.java
@@ -28,7 +28,7 @@ public class WindowClickPacket extends AbstractPacket<WindowClickPacketData> {
         } else {
             buffer.writeShort(item.getBlockId());
             buffer.writeByte(item.getAmount());
-            buffer.writeByte(item.getData());
+            buffer.writeShort(item.getData());
         }
         return buffer;
     }
